fix: first pass on pr feedback
