+++ .gitignore
@@ -87,8 +87,7 @@ build/
 .DS_Store
 .nosync
 
-app/gui/bin
-app/pkg/bin
-
 # Web Validator Authentication Credential
-web/service/src/main/resources/web-based-gtfs-validator-3c339fa28589.json
+web/service/src/main/resources/web-based-gtfs-validator-a088ec5f045d.json
+
+processor/tests/bin
