+++ docs/report.html
@@ -99,7 +99,7 @@
     </style>
 </head>
 
-<body>
+<googleCloudPubsubMessage>
 <h1>GTFS Schedule Validation Report</h1>
 
 
@@ -46031,7 +46031,7 @@
 <footer class="footer text-center text-muted mt-5">
     Made with ♥ by <a href="https://mobilitydata.org/">MobilityData</a>
 </footer>
-</body>
+</googleCloudPubsubMessage>
 <script>
     $(function(){
         $(".accordion tr.notice").on("click", function(){
