--- web/service/src/main/java/org/mobilitydata/gtfsvalidator/web/service/UploadFileResponse.java
@@ -1,10 +1,0 @@
-package org.mobilitydata.gtfsvalidator.web.service;
-
-public class UploadFileResponse {
-
-  private String jobId;
-
-  public UploadFileResponse(String jobId) {
-    this.jobId = jobId;
-  }
-}
