+++ web/service/src/main/java/org/mobilitydata/gtfsvalidator/web/service/controller/CreateJobBody.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Google LLC
+ * Copyright 2023 Jarvus Innovations LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.mobilitydata.gtfsvalidator.web.service;
+package org.mobilitydata.gtfsvalidator.web.service.controller;
 
 public class CreateJobBody {
