+++ web/service/src/main/resources/application.properties
@@ -1,3 +1,3 @@
 spring.cloud.gcp.project-id=web-based-gtfs-validator
-spring.cloud.gcp.credentials.location=classpath:web-based-gtfs-validator-3c339fa28589.json
+spring.cloud.gcp.credentials.location=classpath:web-based-gtfs-validator-a088ec5f045d.json
 logging.level.org.springframework.web=INFO
